requirements.txt
python-dotenv
streamlit
google-genai
fpdf2